+++ client/src/views/dumps/dumps.ts
@@ -1,17 +1,38 @@
-import { Dump, Feed } from "abap-adt-api";
-import { TreeDataProvider, TreeItem, TreeItemCollapsibleState, ViewColumn, window } from "vscode";
-import { getClient } from "../../adt/conections";
-import { AbapFsCommands, command } from "../../commands";
-import { connectedRoots } from "../../config";
+import { Dump, Feed } from "abap-adt-api"
+import { EventEmitter, Range, TreeDataProvider, TreeItem, TreeItemCollapsibleState, Uri, ViewColumn, window, workspace } from "vscode"
+import { getClient } from "../../adt/conections"
+import { AdtObjectFinder } from "../../adt/operations/AdtObjectFinder"
+import { AbapFsCommands, command } from "../../commands"
+import { connectedRoots } from "../../config"
 // tslint:disable:max-classes-per-file
 
+const jsFooter = `<script type="text/javascript">
+const vscode = acquireVsCodeApi();
+const as = document.querySelectorAll("a")
+console.log(as.length)
+as.forEach(
+    a=>a.addEventListener('click',e=>{
+        const uri = e.currentTarget.attributes.href.value
+        if(!uri.match(/^#/)){
+            e.preventDefault();
+            vscode.postMessage({
+                command: 'click',
+                uri
+            });
+        }
+    })
+)</script>`
+
+const inject = (x: string) => `${x}${jsFooter}`
 
 class DumpItem extends TreeItem {
     readonly tag: "dump"
-    dump: Dump;
+    private dump: Dump
+    private connId: string
     constructor(dump: Dump, connId: string) {
         const label = dump.categories.find(c => c.label === "ABAP runtime error")?.term || "dump"
         super(label, TreeItemCollapsibleState.None)
+        this.connId = connId
         this.tag = "dump"
         this.dump = dump
         this.command = {
@@ -22,15 +43,25 @@ class DumpItem extends TreeItem {
     }
     @command(AbapFsCommands.showDump)
     private static show(i: DumpItem) {
-        const panel = window.createWebviewPanel("DUMP", "ABAP Dump", ViewColumn.Active)
-        panel.webview.html = i.dump.text
+        const panel = window.createWebviewPanel("DUMP", "ABAP Dump", ViewColumn.Active, {
+            enableScripts: true, enableCommandUris: true, enableFindWidget: true, retainContextWhenHidden: true
+        })
+        panel.webview.onDidReceiveMessage(async m => {
+            return new AdtObjectFinder(i.connId).displayAdtUri(m.uri)
+        })
+        panel.webview.html = inject(i.dump.text)
     }
 }
 
 class SystemItem extends TreeItem {
     readonly tag = "system"
     private dumpFeed?: Feed | "none"
+    contextValue = "system"
     constructor(label: string, private connId: string) { super(label, TreeItemCollapsibleState.Expanded) }
+    @command(AbapFsCommands.refreshDumps)
+    async refresh(node: any) {
+        dumpProvider.emitter.fire(node)
+    }
     async children() {
         const client = getClient(this.connId)
         if (!this.dumpFeed) {
@@ -46,6 +77,10 @@ class SystemItem extends TreeItem {
 type Item = SystemItem | DumpItem
 class DumpProvider implements TreeDataProvider<Item>{
     private systems = new Map<string, SystemItem>();
+    emitter = new EventEmitter<Item>()
+    get onDidChangeTreeData() {
+        return this.emitter.event
+    }
     getTreeItem(e: Item) {
         return e
     }